fix: change data type

+++ lostandfound/src/main/java/com/example/lostandfound/repository/UserRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
 import com.example.lostandfound.model.User;
 
 @Repository
@@ -13,7 +14,7 @@ public interface UserRepository extends JpaRepository<User, Integer> {
     User findByEmail(String email);
 
     @Query("SELECT u FROM User u WHERE u.id = :id")
-    Optional<User> findById(@Param("id") Long id);
+    Optional<User> findById(@Param("id") Integer id);
 
     // Additional custom queries can be added here
 }
