+++ lostandfound/src/main/java/com/example/lostandfound/service/AdminService.java
@@ -1,13 +1,14 @@
 package com.example.lostandfound.service;
 
+import java.util.List;
+
 import com.example.lostandfound.model.AdminAction;
 import com.example.lostandfound.model.Item;
 import com.example.lostandfound.model.User;
-import java.util.List;
 
 public interface AdminService {
     List<User> getAllUsers();
     List<Item> getAllItems();
-    void approveItemReturn(Long itemId, Long userId);
+    void approveItemReturn(Long itemId, Integer userId);
     AdminAction logAdminAction(AdminAction action);
 }
